admin: factor out security routes
