+++ src/v/redpanda/admin/server.cc
@@ -19,7 +19,6 @@
 #include "cluster/members_frontend.h"
 #include "cluster/metadata_cache.h"
 #include "cluster/partition_manager.h"
-#include "cluster/security_frontend.h"
 #include "cluster/shard_table.h"
 #include "cluster/topics_frontend.h"
 #include "cluster/types.h"
@@ -33,10 +32,7 @@
 #include "redpanda/admin/api-doc/cluster_config.json.h"
 #include "redpanda/admin/api-doc/config.json.h"
 #include "redpanda/admin/api-doc/raft.json.h"
-#include "redpanda/admin/api-doc/security.json.h"
 #include "redpanda/admin/api-doc/status.json.h"
-#include "security/scram_algorithm.h"
-#include "security/scram_authenticator.h"
 #include "vlog.h"
 
 #include <seastar/core/coroutine.hh>
@@ -881,113 +877,6 @@ void admin_server::register_raft_routes() {
                         ss::json::json_void());
                   });
             });
-      });
-}
-
-// TODO: factor out generic serialization from seastar http exceptions
-static security::scram_credential
-parse_scram_credential(const rapidjson::Document& doc) {
-    if (!doc.IsObject()) {
-        throw ss::httpd::bad_request_exception(fmt::format("Not an object"));
-    }
-
-    if (!doc.HasMember("algorithm") || !doc["algorithm"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String algo missing"));
-    }
-    const auto algorithm = std::string_view(
-      doc["algorithm"].GetString(), doc["algorithm"].GetStringLength());
-
-    if (!doc.HasMember("password") || !doc["password"].IsString()) {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("String password smissing"));
-    }
-    const auto password = doc["password"].GetString();
-
-    security::scram_credential credential;
-
-    if (algorithm == security::scram_sha256_authenticator::name) {
-        credential = security::scram_sha256::make_credentials(
-          password, security::scram_sha256::min_iterations);
-
-    } else if (algorithm == security::scram_sha512_authenticator::name) {
-        credential = security::scram_sha512::make_credentials(
-          password, security::scram_sha512::min_iterations);
-
-    } else {
-        throw ss::httpd::bad_request_exception(
-          fmt::format("Unknown scram algorithm: {}", algorithm));
-    }
-
-    return credential;
-}
-
-void admin_server::register_security_routes() {
-    ss::httpd::security_json::create_user.set(
-      _server._routes,
-      [this](std::unique_ptr<ss::httpd::request> req)
-        -> ss::future<ss::json::json_return_type> {
-          auto doc = parse_json_body(*req);
-
-          auto credential = parse_scram_credential(doc);
-
-          if (!doc.HasMember("username") || !doc["username"].IsString()) {
-              throw ss::httpd::bad_request_exception(
-                fmt::format("String username missing"));
-          }
-
-          auto username = security::credential_user(
-            doc["username"].GetString());
-
-          auto err
-            = co_await _controller->get_security_frontend().local().create_user(
-              username, credential, model::timeout_clock::now() + 5s);
-          vlog(logger.debug, "Creating user {}:{}", err, err.message());
-          co_await throw_on_error(*req, err, model::controller_ntp);
-          co_return ss::json::json_return_type(ss::json::json_void());
-      });
-
-    ss::httpd::security_json::delete_user.set(
-      _server._routes,
-      [this](std::unique_ptr<ss::httpd::request> req)
-        -> ss::future<ss::json::json_return_type> {
-          auto user = security::credential_user(req->param["user"]);
-
-          auto err
-            = co_await _controller->get_security_frontend().local().delete_user(
-              user, model::timeout_clock::now() + 5s);
-          vlog(logger.debug, "Deleting user {}:{}", err, err.message());
-          co_await throw_on_error(*req, err, model::controller_ntp);
-          co_return ss::json::json_return_type(ss::json::json_void());
-      });
-
-    ss::httpd::security_json::update_user.set(
-      _server._routes,
-      [this](std::unique_ptr<ss::httpd::request> req)
-        -> ss::future<ss::json::json_return_type> {
-          auto user = security::credential_user(req->param["user"]);
-
-          auto doc = parse_json_body(*req);
-
-          auto credential = parse_scram_credential(doc);
-
-          auto err
-            = co_await _controller->get_security_frontend().local().update_user(
-              user, credential, model::timeout_clock::now() + 5s);
-          vlog(logger.debug, "Updating user {}:{}", err, err.message());
-          co_await throw_on_error(*req, err, model::controller_ntp);
-          co_return ss::json::json_return_type(ss::json::json_void());
-      });
-
-    ss::httpd::security_json::list_users.set(
-      _server._routes, [this](std::unique_ptr<ss::httpd::request>) {
-          std::vector<ss::sstring> users;
-          for (const auto& [user, _] :
-               _controller->get_credential_store().local()) {
-              users.push_back(user());
-          }
-          return ss::make_ready_future<ss::json::json_return_type>(
-            std::move(users));
       });
 }
